requirements.txt
# Core (always required)
pillow>=10.0
numpy>=1.24
pydantic>=2.5
pyyaml>=6.0
tqdm>=4.66
click>=8.1
rich>=13.7

# GPU / model dependencies. Install with `pip install -e .[gpu]` when you have
# a GPU. Pinned wide to keep LLaVA-1.6 (7B) and Mistral-7B reproducible.
# torch>=2.2
# transformers>=4.40
# accelerate>=0.27
# bitsandbytes>=0.43  # skip on macOS
# peft>=0.10
# sentencepiece>=0.2
# einops>=0.7

# Dataset dependencies
# datasets>=2.18
# huggingface_hub>=0.22
# requests>=2.31
